+++ spotless-changelog-plugin-gradle/src/main/java/com/diffplug/spotless/changelog/gradle/ChangelogPlugin.java
@@ -25,15 +25,24 @@ import java.io.IOException;
 import java.nio.charset.StandardCharsets;
 import java.nio.file.Files;
 import java.time.LocalDate;
+import java.util.Collections;
+import java.util.HashSet;
 import java.util.LinkedHashMap;
+import java.util.Set;
 import javax.inject.Inject;
 import org.eclipse.jgit.api.errors.GitAPIException;
 import org.gradle.api.DefaultTask;
 import org.gradle.api.GradleException;
 import org.gradle.api.Plugin;
 import org.gradle.api.Project;
+import org.gradle.api.Task;
 import org.gradle.api.plugins.BasePlugin;
 import org.gradle.api.plugins.JavaBasePlugin;
+import org.gradle.api.provider.Property;
+import org.gradle.api.provider.Provider;
+import org.gradle.api.services.BuildService;
+import org.gradle.api.services.BuildServiceParameters;
+import org.gradle.api.tasks.Internal;
 import org.gradle.api.tasks.TaskAction;
 import org.gradle.api.tasks.TaskProvider;
 
@@ -46,37 +55,90 @@ public class ChangelogPlugin implements Plugin<Project> {
 
 		ChangelogExtension extension = project.getExtensions().create(ChangelogExtension.NAME, ChangelogExtension.class, project);
 		project.getTasks().register(PrintTask.NAME, PrintTask.class, extension);
-		if (project.getRootProject().hasProperty("sshStrictHostKeyChecking")) {
-			extension.gitCfg.sshStrictHostKeyChecking = (String) project.getRootProject().findProperty("sshStrictHostKeyChecking");
+		String sshStrictHostKeyChecking = (String) project.getRootProject().findProperty("sshStrictHostKeyChecking");
+		if (sshStrictHostKeyChecking != null) {
+			extension.data.gitCfg.sshStrictHostKeyChecking = sshStrictHostKeyChecking;
 		}
 
+		Provider<TaskOrderingService> taskOrdering = project.getGradle().getSharedServices().registerIfAbsent("ChangelogService", TaskOrderingService.class, unused -> {});
+
+		TaskProvider<PushWillRunTask> pushWillRun = project.getTasks().register(PushWillRunTask.NAME, PushWillRunTask.class, task -> {
+			task.getTaskOrderingService().set(taskOrdering);
+		});
+
 		TaskProvider<CheckTask> check = project.getTasks().register(CheckTask.NAME, CheckTask.class, extension);
+		check.configure(t -> {
+			t.getTaskOrderingService().set(taskOrdering);
+			t.mustRunAfter(pushWillRun);
+		});
 		TaskProvider<BumpTask> bump = project.getTasks().register(BumpTask.NAME, BumpTask.class, extension);
 		bump.configure(t -> t.dependsOn(check));
 		TaskProvider<PushTask> push = project.getTasks().register(PushTask.NAME, PushTask.class, extension);
-		push.configure(t -> t.dependsOn(bump));
+		push.configure(t -> {
+			t.dependsOn(bump);
+			t.dependsOn(pushWillRun);
+		});
 
 		project.afterEvaluate(unused -> {
-			if (extension.enforceCheck) {
+			if (extension.data.enforceCheck) {
 				project.getTasks().named(JavaBasePlugin.CHECK_TASK_NAME).configure(t -> t.dependsOn(check));
 			}
 		});
 	}
 
+	/**
+	 * +---------------------------+                                        +-------+
+	 * | pushWillRun               |<--dependsOn----------------------------| push  |
+	 * |                           |                                        +---T---+
+	 * | marks that push will run  |                                        dependsOn
+	 * |  in TaskOrderingService   |                 +-------+              +---V---+
+	 * | as a taskGraph workaround |<--mustRunAfter--| check |<--dependsOn--| bump  |
+	 * +---------------------------+                 +-------+              +-------+
+	 */
+	public static abstract class TaskOrderingService implements BuildService<BuildServiceParameters.None> {
+		private final Set<String> pushWillRun = Collections.synchronizedSet(new HashSet<>());
+
+		private String projectFor(Task task) {
+			String path = task.getPath();
+			int lastColon = path.lastIndexOf(':');
+			return lastColon == -1 ? path : path.substring(0, lastColon);
+		}
+
+		void notifyPushWillRun(PushWillRunTask task) {
+			pushWillRun.add(projectFor(task));
+		}
+
+		boolean pushWillRun(CheckTask task) {
+			return pushWillRun.contains(projectFor(task));
+		}
+	}
+
+	public static abstract class PushWillRunTask extends DefaultTask {
+		public static final String NAME = "changelogInternalPushWillRun";
+
+		@Internal
+		abstract Property<TaskOrderingService> getTaskOrderingService();
+
+		@TaskAction
+		public void pushWillRun() {
+			getTaskOrderingService().get().notifyPushWillRun(this);
+		}
+	}
+
 	private static abstract class ChangelogTask extends DefaultTask {
-		protected final ChangelogExtension extension;
+		protected final ChangelogExtension.Data data;
 
 		protected ChangelogTask(ChangelogExtension extension) {
-			this.extension = extension;
+			this.data = extension.data;
 			setGroup("changelog");
 		}
 
 		protected void assertNotSnapshot() {
-			if (extension.getVersionNext().endsWith(ChangelogAndNext.DASH_SNAPSHOT)) {
-				if (extension.nextVersionCfg.appendSnapshot) {
-					throw new GradleException("You must add `-Prelease=true` to remove the -SNAPSHOT from " + extension.getVersionNext());
+			if (data.getVersionNext().endsWith(ChangelogAndNext.DASH_SNAPSHOT)) {
+				if (data.nextVersionCfg.appendSnapshot) {
+					throw new GradleException("You must add `-Prelease=true` to remove the -SNAPSHOT from " + data.getVersionNext());
 				} else {
-					throw new GradleException("It doesn't make sense to put a -SNAPSHOT version into the changelog, nor to make a git tag " + extension.gitCfg.tagPrefix + extension.getVersionNext());
+					throw new GradleException("It doesn't make sense to put a -SNAPSHOT version into the changelog, nor to make a git tag " + data.gitCfg.tagPrefix + data.getVersionNext());
 				}
 			}
 		}
@@ -86,6 +148,9 @@ public class ChangelogPlugin implements Plugin<Project> {
 	public static abstract class CheckTask extends ChangelogTask {
 		public static final String NAME = "changelogCheck";
 
+		@Internal
+		abstract Property<TaskOrderingService> getTaskOrderingService();
+
 		@Inject
 		public CheckTask(ChangelogExtension extension) {
 			super(extension);
@@ -94,29 +159,27 @@ public class ChangelogPlugin implements Plugin<Project> {
 
 		@TaskAction
 		public void check() throws IOException, GitAPIException {
-			String pushTaskPath = getProject().absoluteProjectPath(PushTask.NAME);
-			// if we're going to push later, let's first make sure that will work
-			if (getProject().getGradle().getTaskGraph().hasTask(pushTaskPath)) {
-				GitActions git = extension.gitCfg.withChangelog(extension.changelogFile, extension.model());
+			boolean pushWillRun = getTaskOrderingService().get().pushWillRun(this);
+			if (pushWillRun) {
+				// if we're going to push later, let's first make sure that will work before we bump and publish
+				GitActions git = data.gitCfg.withChangelog(data.changelogFile, data.model());
 				git.assertNoTag();
 				git.checkCanPush();
 			}
 
-			LinkedHashMap<Integer, String> errors = extension.model().changelog().errors();
+			LinkedHashMap<Integer, String> errors = data.model().changelog().errors();
 			if (errors.isEmpty()) {
 				return;
 			}
 
-			String path = getProject().getRootDir().toPath().relativize(extension.changelogFile.toPath()).toString();
-			String allErrors = StringPrinter.buildString(printer -> {
-				errors.forEach((idx, error) -> {
-					if (idx == -1) {
-						printer.println(path + ": " + error);
-					} else {
-						printer.println(path + ":" + idx + ": " + error);
-					}
-				});
-			});
+			String path = data.projectRoot.toPath().relativize(data.changelogFile.toPath()).toString();
+			String allErrors = StringPrinter.buildString(printer -> errors.forEach((idx, error) -> {
+				if (idx == -1) {
+					printer.println(path + ": " + error);
+				} else {
+					printer.println(path + ":" + idx + ": " + error);
+				}
+			}));
 			throw new GradleException(allErrors);
 		}
 	}
@@ -133,10 +196,10 @@ public class ChangelogPlugin implements Plugin<Project> {
 
 		@TaskAction
 		public void print() {
-			if (extension.getVersionNext().equals(extension.getVersionLast())) {
-				System.out.println(getProject().getName() + " " + extension.getVersionLast() + " (no unreleased changes)");
+			if (data.getVersionNext().equals(data.getVersionLast())) {
+				System.out.println(data.projectName + " " + data.getVersionLast() + " (no unreleased changes)");
 			} else {
-				System.out.println(getProject().getName() + " " + extension.getVersionLast() + " -> " + extension.getVersionNext());
+				System.out.println(data.projectName + " " + data.getVersionLast() + " -> " + data.getVersionNext());
 			}
 		}
 	}
@@ -154,15 +217,15 @@ public class ChangelogPlugin implements Plugin<Project> {
 		@TaskAction
 		public void bump() throws IOException {
 			assertNotSnapshot();
-			if (extension.getVersionNext().equals(extension.getVersionLast())) {
+			if (data.getVersionNext().equals(data.getVersionLast())) {
 				// if there are no unreleased changes, then the changelog on disk has already been bumped
 				return;
 			}
 			// time to bump!
-			ChangelogAndNext model = extension.model();
+			ChangelogAndNext model = data.model();
 			LocalDate localDate = LocalDate.now(Time.clockUtc());
 			Changelog bumped = model.changelog().releaseUnreleased(model.versions().next(), localDate.toString());
-			Files.write(extension.changelogFile.toPath(), bumped.toString().getBytes(StandardCharsets.UTF_8));
+			Files.write(data.changelogFile.toPath(), bumped.toString().getBytes(StandardCharsets.UTF_8));
 		}
 	}
 
@@ -179,7 +242,7 @@ public class ChangelogPlugin implements Plugin<Project> {
 		@TaskAction
 		public void push() throws IOException, GitAPIException {
 			assertNotSnapshot();
-			GitActions git = extension.gitCfg.withChangelog(extension.changelogFile, extension.model());
+			GitActions git = data.gitCfg.withChangelog(data.changelogFile, data.model());
 			git.addAndCommit();
 			git.tagBranchPush();
 		}