+++ CHANGELOG.md
@@ -5,6 +5,10 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Added
+- Full support for Gradle Configuration Cache ([#27](https://github.com/diffplug/spotless-changelog/pull/27)).
+### Changed
+- Minimum required Gradle bumped from `5.2` to `6.2` because we need `BuildService` and `ProviderFactory.gradleProperty`.
 
 ## [2.2.0] - 2021-05-11
 ### Added